requirements-test.txt
mypy==0.790
black==19.10b0
pytest==6.1.1
tensorflow-cpu==2.4.0
pytorch-lightning==1.1.1
torch==1.4.0
torchvision==0.5.0
# for integration tests
boto3==1.12.32
google-cloud-storage==1.32.0
azure-storage-blob==2.1.0
waiting==1.4.1
python-dateutil==2.1
